Fix alphashape issue when it fails
Not sure why this started happening only now, or if it will change any results from the paper. But the issue is that when alphashape fails, it fefaults to an alpha value of 0, but this causes an error a few lines later in its code. That's a bug. I had to change this code so that if this error occurs, just manually input an alpha value of 0, which returns the convex hull.